+++ pangeo_forge_recipes/patterns.py
@@ -2,8 +2,9 @@
 Filename / URL patterns.
 """
 import inspect
+import warnings
 from dataclasses import dataclass, field, replace
-from enum import Enum
+from enum import Enum, auto
 from itertools import product
 from typing import (
     Any,
@@ -99,6 +100,21 @@ CombineDim = Union[MergeDim, ConcatDim]
 FilePatternIndex = Index
 
 
+class AutoName(Enum):
+    # Recommended by official Python docs for auto naming:
+    # https://docs.python.org/3/library/enum.html#using-automatic-values
+    def _generate_next_value_(name, start, count, last_values):
+        return name
+
+
+class FileType(AutoName):
+    unknown = auto()
+    netcdf3 = auto()
+    netcdf4 = auto()
+    grib = auto()
+    opendap = auto()
+
+
 class FilePattern:
     """Represents an n-dimensional matrix of individual files to be combined
     through a combination of merge and concat operations. Each operation generates
@@ -114,8 +130,9 @@ class FilePattern:
       May include ``block_size``, ``username``, ``password``, etc.
     :param query_string_secrets: If provided, these key/value pairs are appended to
       the query string of each ``file_pattern`` url at runtime.
-    :param is_opendap: If True, assume all input fnames represent opendap endpoints.
-      Cannot be used with caching.
+    :param file_type: The file format of the source files for this pattern. Must be one of
+      the options defined by ``pangeo_forge_recipes.patterns.FileType``.
+      Note: ``FileType.opendap`` cannot be used with caching.
     """
 
     def __init__(
@@ -124,17 +141,34 @@ class FilePattern:
         *combine_dims: CombineDim,
         fsspec_open_kwargs: Optional[Dict[str, Any]] = None,
         query_string_secrets: Optional[Dict[str, str]] = None,
-        is_opendap: bool = False,
+        file_type: str = "netcdf4",
+        is_opendap: Optional[bool] = None,
     ):
         self.format_function = format_function
         self.combine_dims = combine_dims
         self.fsspec_open_kwargs = fsspec_open_kwargs if fsspec_open_kwargs else {}
         self.query_string_secrets = query_string_secrets if query_string_secrets else {}
+        self.file_type = FileType(file_type)
+
         self.is_opendap = is_opendap
-        if self.fsspec_open_kwargs and self.is_opendap:
+        if self.is_opendap:
+            _deprecation_message = (
+                "`FilePattern(..., is_opendap=True)` will be deprecated in v0.9.0. "
+                "Please use `FilePattern(..., file_type='opendap')` instead."
+            )
+            warnings.warn(_deprecation_message, DeprecationWarning)
+            _maybe_default = "default" if self.file_type.value == "netcdf4" else ""
+            _overide_warning = (
+                f"`is_opendap` passed as `True`, overriding {_maybe_default} "
+                f"`file_type.value == '{self.file_type.value}' with `'opendap'`."
+            )
+            warnings.warn(_overide_warning)
+            self.file_type = FileType("opendap")
+
+        if self.fsspec_open_kwargs and self.file_type == FileType.opendap:
             raise ValueError(
                 "OPeNDAP inputs are not opened with `fsspec`. "
-                "`is_opendap` must be `False` when passing `fsspec_open_kwargs`."
+                "When passing `fsspec_open_kwargs`, `file_type` cannot be `opendap`."
             )
 
     def __repr__(self):