+++ docs/development/release_notes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## v0.8.3 -
+
+- Added `.file_type` attribute to {class}`pangeo_forge_recipes.patterns.FilePattern`. This attribute will eventually supercede
+`.is_opendap`, which will be deprecated in `0.9.0`. Until then, `FilePattern(..., is_opendap=True)` is supported as equivalent
+to `FilePattern(..., file_type="opendap")`. {pull}`322`
+
 ## v0.8.2 - 2022-02-23
 
 - Removed click from dependencies and removed cli entrypoint.